Renames wasm module from "http" to "http-handler" (#6)
This follows conventions similar to WASI where the ABI isn't named based
on which side (host vs guest), but rather which functionality it is
implementing. In this case, we should dodge "http" as it will lead to
confusion later when things like HTTP client ABIs start becoming
defined. "http-handler" is a nice way for us to describe a synchronous
server handler.

https://github.com/WebAssembly/wasi-filesystem/issues/46#issuecomment-1143873563

Signed-off-by: Adrian Cole <adrian@tetrate.io>